Input now available Resolves #18 and closes therefore #2
