+++ app/src/main/java/com/example/repairtool/repairs/Building.kt
@@ -1,0 +1,5 @@
+package com.example.repairtool.repairs
+
+data class Building(val name:String) {
+    override fun toString() = name
+}
