+++ app/src/main/java/com/example/repairtool/repairs/BuildingList.kt
@@ -1,0 +1,12 @@
+package com.example.repairtool.repairs
+
+object BuildingList {
+    var buildingList = mutableListOf(
+        Building(""),
+        Building("Damestoilet"),
+        Building("Hal"),
+        Building("Herentoilet"),
+        Building("Hoofdgebouw"),
+        Building("Kantoor")
+    )
+}
