+++ app/src/main/java/com/example/repairtool/repairs/Repair.kt
@@ -1,3 +1,9 @@
 package com.example.repairtool.repairs
 
-data class Repair(val name: String, val priority: String, val description: String, val status: String)
+data class Repair(
+    val name: String,
+    val building: String,
+    val priority: String,
+    val description: String,
+    val status: String
+    )
