+++ app/src/main/java/com/example/repairtool/repairs/RepairList.kt
@@ -2,8 +2,8 @@ package com.example.repairtool.repairs
 
 object RepairList {
     var repairList = mutableListOf(
-        Repair("Repair 1", "Hoog", "", "Registered"),
-        Repair("Repair 2", "Middel", "", "Registered"),
-        Repair("Repair 3", "Laag", "", "Pending")
+        Repair("Repair 1", "Hoofdgebouw","Hoog", "", "Registered"),
+        Repair("Repair 2", "Hal", "Middel", "", "Registered"),
+        Repair("Repair 3", "Herentoilet","Laag", "", "Pending")
     )
 }
